Initial checkin, but still no where near working yet.
This is the result of several hours of refactoring properJavaRDP to align w/ Android.
Mostly this involved:
1) Removing all references to java.awt
2) Refactoring the use of Bitmap, Canvas, View, etc...
3) Changing the way Keyboard and Mouse events are handled

I did opt to remove the "BufferedBitmap" caching option from the original code.
I am focusing on writing this app to compliment the Android's capabilities.
I am not trying to provide all of the features of properJavaRDP.

RDesktop is hard coded w/ some parameters, but these will be eventually removed and replaced with options in the Connectoid UI.